RATEST-18: Refactor pages from automated test folder
